Fix warnings and missing include, update argument parsing code

+++ src/io_png.c
@@ -674,7 +674,7 @@ static int io_png_write_raw(const char *fname, const void *data,
                 /* row loop */
                 for (i = 0; i < nx; i++) {
                     /* pixel loop */
-                    tmp = floor(*data_f32_ptr++ + .5f);
+                    tmp = (float)floor(*data_f32_ptr++ + .5f);
                     *idata_ptr = (png_byte) (tmp < 0. ? 0. :
                                              (tmp > 255. ? 255. : tmp));
                     idata_ptr += nc;
